[WIP] Isolate kernel compute graph from task graph
In order to let the runtime have more flexibility to schedule the
execution of streaming applications, this commit introduces the concept
of task, which the schedule module creates dynamically for kernels.
Different types of tasks are allowed, for example, polling worker tasks
keep polling on the message queue, or one-shot tasks that are created
by the data producer when the produced data satisfies the condition to
fire a kernel.

By this commit, only the very basic tests are running properly, such as
bracketsTest, armqMini. Some known issues are
 - some non-trivial FIFO interfaces are not available like allocate()
 with argument pack, allocate_s()
 - the StreamingData approach currently assume the kernel compute()
 would always take one piece of data and produce one piece of data per
 port, which significantly limits the use case
 - lambdak that creates a kernel from a lambda function is not available
 - non-string port name might be broken
 - split, join semantics are dropped
 - discard kernel CLONE() as well
 - there are probably memory leakage here and there, especially on newed
 Kpair and StreamingData objects
 - task id should be used to index some bookkeeping structures, rather
 than using the task pointer to avoid potential ABA problems