[WIP] Add and correct some streaming operators
There was a bug that use Kpair's src_name as destination name (when the
Kpair appears on the right hand-size of >> operator) or dst_name as
source name by mistake. This commit fixs this bug as well as adds a few
missing streaming operators.

This commit also updates the test for StreamingData with the respect to
the method name changes (get -> getT, set -> setT for disambiguation )
in an earlier commit.