[WIP] Allocate buffer for oneshot tasks on demand
This commit makes to change to avoid oneshot tasks from sharing one
ringbuffer (could cause race condition as the underlying FIFO structure
is designed for single producer, single consumer). Instead, we allocate
the output buffer for each oneshot task.
Note: there is obviously memory leakage on the dynamically allocated
buffer, and it is also not very efficient to invoke malloc frequently
and not control the alginments (leads to false sharing).