[WIP] HighwayHash port name at compile time
This commit brings back the support for non-string name port. When
STRING_NAMES is not defined or set to 0, compiler would execute
HighwayHash algorithm to hash the string literals used for port names
(suffixed with `_port`), converting the string into a 64-bit hash key
and encode it into the binary. This way, the expensive and inefficient
STL hash function for string is avoided.