[WIP] Merge SchedMeta into task
The Task and its subclasses are supposed to hold all the context meta
data related to the task execution. The subclasses of Task are crafted
for different schedulers, so it makes no sense to have a pointer
pointing to the meta data customized by the scheduler. Instead, the
scheduler could just define the meta data in the Task subclasses.
Other than making the design more logical, it should have performance
benefits when using std::thread and QThread. Because those two threading
options has the main thread monitoring a linked list of tasks. If the
scheduler meta data is stored nearby the flag that the main thread is
checking, there would be false sharing.