[WIP] Move ready-only data into kernel
This commits cleans up the data structures used by AllocateFIFO.
KernelFIFOAllocMeta holds the read-only data that would not change
during the execution.
When multiple FIFOs exists for a port, RRTaskFIFOAllocMeta is used to
save the offset for the FIFO array. RRTaskFIFOAllocMeta embeds the
Round-Robin algorithm to iterate the FIFOs of a port.
StreamingData now tasks the responsibility to memorize which port is
selected, so that OneShotTask can keep using the per-kernel alloc meta
without worrying about selecting from multiple ports.
PollingWorkerCV is renamed to CondVarWorker and will always use
RRTaskFIFOAllocMeta to save the consumer information.
This commits also depracates some Allocate methods not needed any more.