[WIP] Fix bug: stale task pointer in reused DataIn
The `out2in1piece()` was designed to simply move `stream_out` of the
producer `OneShotTask` to be `stream_in` of the new consumer
`OneShotTask` so that we can save the effort to construct a new
StreamingData structure. However, after `stream_out` is handed over to
the new task, the task pointer in StreamingData is left unchanged, and
also the producer task is still dangerously keeping the `stream_out`
pointer and touch it later in `taskCommit()`. This could lead to
undefined behaviors (segfault occasionally). This commit correct those
bugs.
This commit also add the missing free for the alloc meta allocated by
AllocateNew.