[WIP] Consumer initialization and wakeup corrected
When there are multiple clones of a kernel, the registered consumer
should not be overwritten, so every CondVarWorker should only register
to its own input FIFOs.
When allocating the consumers array, make sure it is initialized to all
zeros.
When invalidating output buffers, check the consumer pointer and get the
correct index when iterating through all output FIFOs this worker owns.