[WIP] Static scheduler task handlers
This commit replaces all task handler functions in every Scheduler class
to their static equivalents in order to avoid the overhead of vtable
look up in the previous virtual method approach. With either converting
some of the methods into templates or apply the static shadowing
technique to selectively "override" some inherited static method, this
commit still trys to reduce the duplication of code as much as I can.
This commit also fixs a minor bug in TaskFIFOAllocMeta that
port_out_selected is not properly initialized, causing segfault.