[WIP] Resizable linked ring buffer
This commit implements a new FIFO, which is a ring buffer formed by
several linked chunk of the same size. Accessing it is similar to ring
buffer on the heap, while the advantage is this ring buffer can be
easily resized without copying. Resizing is done by the producer when it
reaches the end of a chunk and finds the next chunk is currently used by
the consumer. In this case, the producer allocates a new chunk and
insert it into the ring buffer.